перестроение файлов проекта согласно инфраструктуре папок MVVM
