+++ WpfMailSender/App.config
@@ -4,7 +4,7 @@
     </configSections>
     <connectionStrings>
         <add name="WpfMailSender.Properties.Settings.MailSenderDBConnectionString"
-            connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\Data\MailSenderDB.mdf;Integrated Security=True"
+            connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\DataBase\MailSenderDB.mdf;Integrated Security=True"
             providerName="System.Data.SqlClient" />
     </connectionStrings>
     <startup>
