--- WpfMailSender/AuthorizationWindow.xaml
@@ -1,47 +1,0 @@
-﻿<Window x:Class="WpfMailSender.AuthorizationWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfMailSender"
-        mc:Ignorable="d"
-        Title="Авторизация" Height="220" Width="350" ResizeMode="NoResize" WindowStartupLocation="CenterOwner">
-    <DockPanel>
-        <StackPanel Orientation="Horizontal"
-                    DockPanel.Dock="Bottom"
-                    Margin="10,0,10,15" HorizontalAlignment="Center" VerticalAlignment="Center">
-            
-            <Button Content="OK"
-                    Width="100" 
-                    HorizontalAlignment="Left" Click="Button_Click"/>
-            
-            <Button Content="Отмена"
-                    Width="100"
-                    Margin="50,0,0,0" Click="Button_Click_1" />
-        </StackPanel>
-
-        <Border Margin="10" Padding="10,5,10,0"
-                BorderBrush="LightBlue" BorderThickness="1" CornerRadius="5">
-
-            <StackPanel>
-                <!--<StackPanel>
-                    <TextBlock Text="Сервер:" />
-                    <ComboBox ItemsSource="{Binding smtpSettings}" SelectedValue="{Binding selectedSmtpServer}">
-                        <ComboBox.ItemTemplate>
-                            <DataTemplate>
-                                <TextBlock Text="{Binding Path=Name}"/>
-                            </DataTemplate>
-                        </ComboBox.ItemTemplate>
-                    </ComboBox>
-                </StackPanel>-->
-
-                <StackPanel >
-                    <TextBlock Text="Логин:"/>
-                    <TextBox Text="{Binding authSettings.EmailFrom, UpdateSourceTrigger=PropertyChanged}"/>
-                    <TextBlock Text="Пароль:" Margin="0,5,0,0"/>
-                    <TextBox Text="{Binding authSettings.Password}"/>
-                </StackPanel>
-            </StackPanel>
-        </Border>
-    </DockPanel>
-</Window>
