--- WpfMailSender/DataBase.dbml
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?><Database Name="MailSenderDB" Class="DataEmailsDataContext" xmlns="http://schemas.microsoft.com/linqtosql/dbml/2007">
-  <Connection Mode="AppSettings" ConnectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\Data\MailSenderDB.mdf;Integrated Security=True" SettingsObjectName="WpfMailSender.Properties.Settings" SettingsPropertyName="MailSenderDBConnectionString" Provider="System.Data.SqlClient" />
-  <Table Name="dbo.Emails" Member="Emails">
-    <Type Name="Emails">
-      <Column Name="Id" Type="System.Int32" DbType="Int NOT NULL" IsPrimaryKey="true" CanBeNull="false" />
-      <Column Name="Name" Type="System.String" DbType="NVarChar(MAX) NOT NULL" CanBeNull="false" />
-      <Column Name="Email" Type="System.String" DbType="NVarChar(MAX) NOT NULL" CanBeNull="false" />
-    </Type>
-  </Table>
-  <Table Name="dbo.Smtp" Member="Smtp">
-    <Type Name="Smtp">
-      <Column Name="Id" Type="System.Int32" DbType="Int NOT NULL IDENTITY" IsPrimaryKey="true" IsDbGenerated="true" CanBeNull="false" />
-      <Column Name="Name" Type="System.String" DbType="NVarChar(MAX) NOT NULL" CanBeNull="false" />
-      <Column Name="Server" Type="System.String" DbType="NVarChar(MAX) NOT NULL" CanBeNull="false" />
-      <Column Name="Port" Type="System.Int32" DbType="Int NOT NULL" CanBeNull="false" />
-    </Type>
-  </Table>
-</Database>
