--- WpfMailSender/DataBase.dbml.layout
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ordesignerObjectsDiagram dslVersion="1.0.0.0" absoluteBounds="0, 0, 11, 8.5" name="DataEmails">
-  <DataContextMoniker Name="/DataEmailsDataContext" />
-  <nestedChildShapes>
-    <classShape Id="3926c485-2733-47ab-b81f-4314853fa583" absoluteBounds="5, 1.75, 2, 1.3862939453125">
-      <DataClassMoniker Name="/DataEmailsDataContext/Emails" />
-      <nestedChildShapes>
-        <elementListCompartment Id="e51d9135-2af9-4446-afc7-0e94a5c5577c" absoluteBounds="5.015, 2.21, 1.9700000000000002, 0.8262939453125" name="DataPropertiesCompartment" titleTextColor="Black" itemTextColor="Black" />
-      </nestedChildShapes>
-    </classShape>
-    <classShape Id="76595476-a682-4023-ae78-7b4261be5fa0" absoluteBounds="5, 3.375, 2, 1.5785953776041666">
-      <DataClassMoniker Name="/DataEmailsDataContext/Smtp" />
-      <nestedChildShapes>
-        <elementListCompartment Id="e944715e-8027-4200-92f5-2646c76497e3" absoluteBounds="5.015, 3.835, 1.9700000000000002, 1.0185953776041665" name="DataPropertiesCompartment" titleTextColor="Black" itemTextColor="Black" />
-      </nestedChildShapes>
-    </classShape>
-  </nestedChildShapes>
-</ordesignerObjectsDiagram>
