--- WpfMailSender/SendEndWindow.xaml.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace WpfMailSender
-{
-    /// <summary>
-    /// Логика взаимодействия для SendEndWindow.xaml
-    /// </summary>
-    public partial class SendEndWindow : Window
-    {
-        public SendEndWindow()
-        {
-            InitializeComponent();
-        }
-
-        public SendEndWindow(string messageText):this()
-        {
-            tbText.Text = messageText;
-        }
-
-        private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            DialogResult = true;
-        }
-    }
-}
