+++ WpfMailSender/ViewModels/Base/ViewModelBase.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.ComponentModel;
+using System.Threading.Tasks;
+using System.Runtime.CompilerServices;
+
+namespace WpfMailSender.ViewModels.Base
+{
+    internal abstract class ViewModelBase : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        protected virtual void OnPropertyChanged([CallerMemberName] string propName = null)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propName));
+        }
+
+        protected virtual bool Set<T>(ref T field, T value, [CallerMemberName] string propName = null)
+        {
+            if (Equals(field, value)) return false;
+            field = value;
+            OnPropertyChanged(propName);
+            return true;
+        }
+    }
+}
