+++ WpfMailSender/Views/AuthorizationWindow.xaml.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+using WpfMailSender.Models;
+
+namespace WpfMailSender
+{
+    /// <summary>
+    /// Логика взаимодействия для AuthorizationWindow.xaml
+    /// </summary>
+    public partial class AuthorizationWindow : Window
+    {
+        public AuthSettings authSettings { get; set; } = new AuthSettings();
+
+        public AuthorizationWindow()
+        {
+            InitializeComponent();
+            DataContext = this;
+        }
+
+        private void Button_Click(object sender, RoutedEventArgs e)
+        {
+            DialogResult = true;
+        }
+
+
+        private void Button_Click_1(object sender, RoutedEventArgs e)
+        {
+            DialogResult = false;
+        }
+    }
+}
