+++ WpfMailSender/Views/SendEndWindow.xaml
@@ -1,0 +1,25 @@
+﻿<Window x:Class="WpfMailSender.SendEndWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:WpfMailSender"
+        mc:Ignorable="d"
+        Title="Информация" Height="220" Width="350" 
+        ResizeMode="NoResize" WindowStartupLocation="CenterOwner">
+    <Grid >
+        <DockPanel >
+            <Button DockPanel.Dock="Bottom"
+                    Content="OK" Width="100" 
+                    Margin="0,0,0,15" Click="Button_Click"/>
+
+            <Border Margin="10" Padding="5"
+                    BorderBrush="LightBlue" BorderThickness="1" CornerRadius="5">
+                <ScrollViewer>
+                    <TextBlock x:Name="tbText" TextAlignment="Center"
+                        TextWrapping="Wrap" FontSize="14"/>
+                </ScrollViewer>
+            </Border>
+        </DockPanel>
+    </Grid>
+</Window>
